separacao da interacao em classes > > Co-authored-by:  <LimirioGuimaraes>
