+++ src/pages/canvas/Interacao.css
@@ -1,0 +1,138 @@
+/*Interação*/
+
+.form_interacao{
+    display: none;
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0,0,0,0.5);
+    font-family: 'Ubuntu';
+  }
+  
+  .center{
+    position:absolute;
+    display: block;
+    justify-content: center;
+    align-items: center;
+    padding: 30px;
+    margin-right: 500px;
+    background-color: #252525;
+    max-width: 1000px;
+    margin: 30px 8%;
+    border-radius: 30px;
+  }
+  
+  .form_interacao .center button{
+    cursor: pointer;
+    width: 20px;
+    height: 20px;
+    background-color: red;
+    border-radius: 10px;
+    position: absolute;
+    border: none;
+    top: 2px;
+  
+  }
+  
+  .parte1_interacao{
+    float: left;
+    width: 20%;
+  }
+  
+  .parte1_interacao p{
+    color: white;
+    font-size: 20px;
+  }
+  
+  .tabela_analise_custos{
+    margin-top: 20px;
+    margin-bottom: 30px;
+    float: left;
+    width: 90%;
+    font-weight: 100;
+  }
+  
+  .tabela_analise_custos table{
+    box-sizing: border-box;
+    left: 8.11%;
+    right: 41.28%;
+    top: 12.75%;
+    
+    text-align: center;
+    background-color: #2666E3;
+    border: 3px solid #2666E3;
+    border-radius: 12px;
+    color: white;
+    
+  }
+  
+  .tabela_analise_custos table td{
+    height: 30px;
+    border-radius: 10px;
+    font-weight: 400;
+  }
+  
+  .tabela_analise_custos table input{
+    width: 97%;
+    height: 96%;
+    background-color: #252525;
+    border: 1px solid #252525;
+    border-radius: 0px;
+    color: white;
+  }
+  
+  /*config tabela 2*/
+  .parte2_interacao{
+    float: left;
+    width: 33%;
+  }
+  
+  .parte2_interacao p{
+    color: white;
+    font-size: 20px;
+  }
+  
+  .tabela_analise_preço{
+    margin-top: 30px;
+    margin-bottom: 10px;
+    float: left;
+    width: 90%;
+    font-weight: 100;
+  }
+  
+  .tabela_analise_preço table{
+    box-sizing: border-box;
+    left: 8.11%;
+    right: 41.28%;
+    top: 12.75%;
+    
+    text-align: center;
+    background-color: #2666E3;
+    border: 3px solid #2666E3;
+    border-radius: 12px;
+    color: white;
+    
+  }
+  
+  .tabela_analise_preço table td{
+    height: 70px;
+    border-radius: 10px;
+    font-weight: 400;
+  }
+  
+  .tabela_analise_preço table input{
+    width: 97%;
+    height: 96%;
+    background-color: #252525;
+    border: 1px solid #252525;
+    border-radius: 0px;
+    color: white;
+  }
+  
+  .right-arrow {
+    float:right; 
+    width: 60px;
+    height: 60px;
+    cursor: pointer;
+  }
+
